make test data generation scripts a compiled executable
