+++ root_numpy/tests/data/.gitignore
@@ -1,0 +1,2 @@
+generate
+dicts.*
