+++ root_numpy/tests/data/LinkDef.h
@@ -1,0 +1,14 @@
+#include <vector>
+#ifdef __CINT__
+
+#pragma link off all globals;
+#pragma link off all classes;
+#pragma link off all functions;
+
+#pragma link C++ class std::vector<float>+ ;
+#pragma link C++ class std::vector<double>+ ;
+#pragma link C++ class std::vector<long>+ ;
+#pragma link C++ class std::vector<char>+ ;
+#pragma link C++ class std::vector<bool>+ ;
+
+#endif
